Made some changes such as the API used and decimal size for the dividend calculations.

+++ src/main/java/com/example/stocks/Record/Dividends.java
@@ -3,13 +3,14 @@ package com.example.stocks.Record;
 import com.example.stocks.Model.DividendCalender;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record Dividends(
-        @JsonProperty("Annual_Dividend") double annualDividend,
-        @JsonProperty("Monthly_Dividend") double monthlyDividend,
-        @JsonProperty("Days_Dividend") double dailyDividend,
-        @JsonProperty("Hourly_Dividend") double hourlyDividend,
-        @JsonProperty("Yield_On_Cost") double yieldOnCost,
+        @JsonProperty("Annual_Dividend") BigDecimal annualDividend,
+        @JsonProperty("Monthly_Dividend") BigDecimal monthlyDividend,
+        @JsonProperty("Days_Dividend") BigDecimal dailyDividend,
+        @JsonProperty("Hourly_Dividend") BigDecimal hourlyDividend,
+        @JsonProperty("Yield_On_Cost") BigDecimal yieldOnCost,
         @JsonProperty("Dividend_Payments")List<DividendCalender> dividendCalenders
 ) {}
